+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 20.02
 
-+ None
++ Change example list and collection list as a function.
++ Add an iterator for grabbing examples and collections names.
 
 # Version 20.01